ref/home_agent/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
aiohttp==3.9.1
beautifulsoup4==4.12.2
openai==1.3.7
litellm==1.10.3
google-generativeai==0.3.1
loguru==0.7.2
youtube-transcript-api==0.6.1
langchain==0.0.350
jinja2==3.1.2
Pillow==10.1.0  # For image processing
python-magic==0.4.27  # For better file type detection
fastapi-cache2
redis
aioredis